fix(test): stabilize preflight under local FS load
Cap preflight unit/coverage workers at 2, retry Darwin renamex, raise
lifecycle upgrade test timeouts, and split rename exclusive into its own
module to stay under the file-size hard cap.
